Make it so that we purge the URL for the post if the service ID hasn't been included.
This is for users who mix and match two different services (e.g static and dynamic content).

It does *not* purge related content. It probably should.

One problem is that get_permalink($post_id) can optionally return the url with or without the post name being included.

I'm not sure whether we:

- can find out programatically which one is being used by this WP install
- should purge both?
- should tell customers doing this that they should remove the name in vcl_hash